third_party/blink/web_tests/fast/forms/calendar-picker/calendar-picker-click-disabled-day.html

<!DOCTYPE html>
<meta name=fuzzy content="maxDifference=0-3; totalPixels=0-1000">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../resources/common.js"></script>
<script src="../resources/picker-common.js"></script>
<script src="../calendar-picker/resources/calendar-picker-common.js"></script>


<input type=date id=date value="2019-09-20" min="2019-09-17" max="2019-09-25">

<script>
let t = async_test('Test clicking on disabled day');

function test1() {
  removeCommitDelay();
  clickDayCellAt(6, 0);
  assert_not_equals(internals.pagePopupWindow, null);
}

t.step(() => {
  openPicker(document.getElementById('date'))
  .then(t.step_func_done(test1));
});
</script>
